fix(join-requests): fix tentative approval flow bugs (Issue 850)
- tentatively approve now opens a welcome-message dialog (sms/whatsapp
  send, no magic link) mirroring the approve flow, instead of no
  follow-up dialog at all
- approve dialog no longer shows an "edit tentative approval message"
  button — it only edits the shared welcome template
- tentative approval message template now supports ${SENDER_NAME}
- restyle the message-template settings row as a labeled, bordered
  section instead of bare underlined text links
- split JoinRequestCard out of JoinRequestsScreen to stay under the
  500-line file limit
