+++ frontend/src/screens/admin/JoinRequestMessageEditors.tsx
@@ -20,34 +20,30 @@ export function JoinRequestMessageEditors() {
   if (!hasPermission(currentUser, Permission.ApproveJoinRequests)) return null;
 
   return (
-    <div className="mb-4 flex flex-wrap gap-x-4 gap-y-1">
-      <button
-        type="button"
-        onClick={() => {
-          setWelcomeOpen(true);
-        }}
-        className="text-muted hover:text-foreground text-xs underline"
-      >
-        edit shared welcome template
-      </button>
-      <button
-        type="button"
-        onClick={() => {
-          setTentativeOpen(true);
-        }}
-        className="text-muted hover:text-foreground text-xs underline"
-      >
-        edit tentative approval message
-      </button>
-      <button
-        type="button"
-        onClick={() => {
-          setWhatsappOpen(true);
-        }}
-        className="text-muted hover:text-foreground text-xs underline"
-      >
-        edit whatsapp link
-      </button>
+    <div className="border-border bg-surface-dim mb-4 rounded-md border p-3">
+      <p className="text-muted mb-2 text-xs font-medium tracking-wide uppercase">
+        message templates
+      </p>
+      <div className="flex flex-wrap gap-2">
+        <EditorTrigger
+          label="edit shared welcome template"
+          onClick={() => {
+            setWelcomeOpen(true);
+          }}
+        />
+        <EditorTrigger
+          label="edit tentative approval message"
+          onClick={() => {
+            setTentativeOpen(true);
+          }}
+        />
+        <EditorTrigger
+          label="edit whatsapp link"
+          onClick={() => {
+            setWhatsappOpen(true);
+          }}
+        />
+      </div>
       <WelcomeTemplateEditorDialog
         open={welcomeOpen}
         onClose={() => {
@@ -70,5 +66,17 @@ export function JoinRequestMessageEditors() {
         whatsappLink={whatsappQ.data ?? null}
       />
     </div>
+  );
+}
+
+function EditorTrigger({ label, onClick }: { label: string; onClick: () => void }) {
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+      className="border-border-strong bg-surface text-foreground-secondary hover:bg-background focus-visible:ring-brand-200 rounded-md border px-2.5 py-1 text-xs font-medium transition-colors focus-visible:ring-2 focus-visible:outline-none"
+    >
+      {label}
+    </button>
   );
 }
