+++ frontend/src/screens/admin/TentativeApprovalMessageDialog.tsx
@@ -1,20 +1,14 @@
 import { useState } from 'react';
 
-import { useWelcomeTemplate, useWhatsAppLink } from '@/api/content';
+import { useTentativeApprovalMessage, useWhatsAppLink } from '@/api/content';
 import { useAuthStore } from '@/auth/store';
 import { Button } from '@/components/ui/Button';
 import { Dialog } from '@/components/ui/Dialog';
 import { hasPermission, Permission } from '@/models/permissions';
 import { formatPhone } from '@/utils/formatPhone';
-import {
-  buildMagicLinkUrl,
-  buildSmsHref,
-  buildWelcomeMessage,
-  buildWhatsAppHref,
-  renderWelcomeMessage,
-} from '@/utils/welcomeMessage';
+import { buildSmsHref, buildWhatsAppHref, renderWelcomeMessage } from '@/utils/welcomeMessage';
 
-import { WelcomeTemplateEditorDialog } from './WelcomeTemplateEditorDialog';
+import { TentativeApprovalMessageEditorDialog } from './TentativeApprovalMessageEditorDialog';
 
 interface Props {
   open: boolean;
@@ -22,62 +16,45 @@ interface Props {
   fullName: string;
   firstName: string;
   phoneNumber: string;
-  magicLinkToken: string | null;
 }
 
-export function ApprovalCredentialsDialog({
+const DEFAULT_TENTATIVE_MESSAGE =
+  "hi ${FIRST_NAME} 🌱 you're tentatively in! come to an event in person and we'll get you fully approved.";
+
+export function TentativeApprovalMessageDialog({
   open,
   onClose,
   fullName,
   firstName,
   phoneNumber,
-  magicLinkToken,
 }: Props) {
-  const [copied, setCopied] = useState(false);
   const [editorOpen, setEditorOpen] = useState(false);
   const currentUser = useAuthStore((s) => s.user);
-  const templateQ = useWelcomeTemplate();
+  const templateQ = useTentativeApprovalMessage();
   const whatsappLinkQ = useWhatsAppLink();
 
-  if (!magicLinkToken) return null;
-  const magicLinkUrl = buildMagicLinkUrl(magicLinkToken);
   const senderName = currentUser?.firstName ?? '';
-  // If the template fetch fails, fall back to the legacy hardcoded body so
-  // vetters can still send a message.
-  const welcomeMessage = templateQ.data
-    ? renderWelcomeMessage(templateQ.data.body, {
-        name: firstName,
-        senderName,
-        magicLink: magicLinkUrl,
-        whatsappLink: whatsappLinkQ.data?.link ?? '',
-      })
-    : buildWelcomeMessage(firstName, magicLinkUrl);
-  const smsHref = buildSmsHref(phoneNumber, welcomeMessage);
-  const whatsappHref = buildWhatsAppHref(phoneNumber, welcomeMessage);
+  const body = templateQ.data?.body.trim() ? templateQ.data.body : DEFAULT_TENTATIVE_MESSAGE;
+  const message = renderWelcomeMessage(body, {
+    name: firstName,
+    senderName,
+    whatsappLink: whatsappLinkQ.data?.link ?? '',
+  });
+  const smsHref = buildSmsHref(phoneNumber, message);
+  const whatsappHref = buildWhatsAppHref(phoneNumber, message);
   const sendButtonsDisabled = templateQ.isPending;
   const canEditTemplate = hasPermission(currentUser, Permission.ApproveJoinRequests);
-
-  async function copyLink() {
-    await navigator.clipboard.writeText(magicLinkUrl);
-    setCopied(true);
-    window.setTimeout(() => {
-      setCopied(false);
-    }, 2000);
-  }
 
   return (
     <>
       <Dialog open={open} onClose={onClose} title={`welcome ${fullName}`}>
         <p className="text-foreground-secondary text-sm">
-          share this one-time login link with {formatPhone(phoneNumber)}. it won't be shown again.
+          let {formatPhone(phoneNumber)} know they're tentatively in.
         </p>
-        <div className="bg-surface-dim mt-3 overflow-x-auto rounded-md px-3 py-2 font-mono text-xs break-all">
-          {magicLinkUrl}
+        <div className="bg-surface-dim mt-3 overflow-x-auto rounded-md px-3 py-2 text-xs break-words whitespace-pre-wrap">
+          {message}
         </div>
         <div className="mt-3 flex flex-wrap gap-2">
-          <Button variant="secondary" onClick={() => void copyLink()}>
-            {copied ? 'copied ✓' : 'copy link'}
-          </Button>
           <SendLink href={smsHref} label="send via sms" disabled={sendButtonsDisabled} />
           <SendLink href={whatsappHref} label="send via whatsapp" disabled={sendButtonsDisabled} />
         </div>
@@ -90,7 +67,7 @@ export function ApprovalCredentialsDialog({
               }}
               className="text-muted hover:text-foreground text-left text-xs underline"
             >
-              edit shared welcome template
+              edit tentative approval message
             </button>
           </div>
         ) : null}
@@ -98,7 +75,7 @@ export function ApprovalCredentialsDialog({
           <Button onClick={onClose}>done</Button>
         </div>
       </Dialog>
-      <WelcomeTemplateEditorDialog
+      <TentativeApprovalMessageEditorDialog
         open={editorOpen}
         onClose={() => {
           setEditorOpen(false);
