+++ frontend/src/screens/admin/TentativeApprovalMessageEditorDialog.tsx
@@ -53,13 +53,14 @@ function EditorForm({ initialBody, onClose }: { initialBody: string; onClose: ()
   return (
     <form onSubmit={(e) => void onSubmit(e)} className="flex flex-col gap-3">
       <p className="text-foreground-secondary text-sm">
-        sent when someone who came in person gets fully approved — this replaces the default
-        confirmation text.
+        sent to someone when they're tentatively approved, and again when they're fully approved
+        after coming in person — this replaces the default confirmation text.
       </p>
       <p className="text-muted text-xs">
         available placeholders:{' '}
         <code className="bg-surface-dim rounded px-1">{'${FIRST_NAME}'}</code> (recipient's first
-        name), <code className="bg-surface-dim rounded px-1">{'${WHATSAPP_LINK}'}</code>
+        name), <code className="bg-surface-dim rounded px-1">{'${SENDER_NAME}'}</code>,{' '}
+        <code className="bg-surface-dim rounded px-1">{'${WHATSAPP_LINK}'}</code>
       </p>
       <textarea
         value={body}
